Refactor SCIM extension registration
Fixes #67
