+++ .github/workflows/runchecks.yml
@@ -4,9 +4,6 @@ on: [push]
 
 jobs:
   build:
-    env:
-      WANT_ETHERNET_MAB: True
-      WANT_FDO: True
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
