+++ gateway/scim_ethermab.py
@@ -12,14 +12,15 @@ from ciscoisesdk.exceptions import ApiError
 from sqlalchemy import ForeignKey, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, relationship, mapped_column
-from scim_extensions import scim_ext_create, scim_ext_read, \
-    scim_ext_update, scim_ext_delete
+from scim_extensions import register_scim_extension
 from models import Device
 from database import session,db
 from config import ISE_SUPPORT, ISE_HOST, ISE_USERNAME, ISE_PASSWORD, \
     WANT_ETHER_MAB
 from tiedie_exceptions import SchemaError, DeviceExists, ISEError, \
     MABNotSupported
+
+ETHER_MAB_ENABLED = bool(WANT_ETHER_MAB)
 
 class EtherMABExtension(db.Model):
     """
@@ -41,7 +42,7 @@ class EtherMABExtension(db.Model):
         Populate Object with the two required attributes.
         """
 
-        if not WANT_ETHER_MAB:
+        if not ETHER_MAB_ENABLED:
             raise MABNotSupported
 
         self.device_id = device_id
@@ -204,11 +205,22 @@ def ethermab_delete_device(entry_id):
                 raise ISEError(e.description) from e
     session.delete(entry)
     session.commit()
+def register_ethermab_extension(enabled=None):
+    """Register ethernet MAB extension hooks when enabled."""
+    # pylint: disable=global-statement
+    global ETHER_MAB_ENABLED
 
+    if enabled is None:
+        ETHER_MAB_ENABLED = bool(WANT_ETHER_MAB)
+    else:
+        ETHER_MAB_ENABLED = bool(enabled)
 
+    if not ETHER_MAB_ENABLED:
+        return
 
-if WANT_ETHER_MAB:
-    scim_ext_create.append(ethermab_create_device)
-    scim_ext_read.append(ethermab_read_device)
-    scim_ext_update.append(ethermab_update_device)
-    scim_ext_delete.append(ethermab_delete_device)
+    register_scim_extension(
+        ethermab_create_device,
+        ethermab_read_device,
+        ethermab_update_device,
+        ethermab_delete_device,
+    )
